parcel create feature

+++ parcel/models.py
@@ -11,7 +11,9 @@ class Parcel(models.Model):
     source_pincode = models.CharField(max_length=10)
     destination_pincode = models.CharField(max_length=10)
 
-    weight = models.FloatField()
+    # Better for precision (instead of FloatField)
+    weight = models.DecimalField(max_digits=5, decimal_places=2)
+
     dimensions = models.CharField(max_length=100, blank=True)
 
     created_at = models.DateTimeField(auto_now_add=True)
